[NewIR]Split NewIRCompiler with .h/.cc and decoupling compilatiom with cinncore
