+++ test/cpp/ir/cinn/CMakeLists.txt
@@ -1,8 +1,8 @@
 if(WITH_TESTING AND WITH_CINN)
   cc_test_old(
-    test_graph_compiler_new_ir
+    test_new_ir_compiler
     SRCS
-    graph_compiler_new_ir_test.cc
+    new_ir_compiler_test.cc
     DEPS
     cinncore
     pd_dialect
@@ -10,6 +10,5 @@ if(WITH_TESTING AND WITH_CINN)
     phi
     gtest
     glog)
-  set_tests_properties(test_graph_compiler_new_ir PROPERTIES LABELS
-                                                             "RUN_TYPE=CINN")
+  set_tests_properties(test_new_ir_compiler PROPERTIES LABELS "RUN_TYPE=CINN")
 endif()
